+++ l10n_nl_base_ubl/__manifest__.py
@@ -2,13 +2,13 @@
 # License AGPL-3.0 or later (https://www.gnu.org/licenses/agpl).
 
 {
-    "name": "Base UBL - Netherlands",
+    "name": "Dutch - UBL",
     "version": "13.0.1.0.0",
     "category": "Dutch Localization",
     "license": "AGPL-3",
-    "summary": "Base module for Netherlands-specific UBL dialect (NLCIUS)",
+    "summary": "Adds Dutch-related data in UBL XML files",
     "author": "Onestein, Odoo Community Association (OCA)",
-    "maintainers": ["astirpe", "thomaspaulb"],
+    "maintainers": ["astirpe"],
     "website": "https://github.com/OCA/l10n-netherlands",
     "depends": ["base_ubl"],
     "installable": True,
