--- l10n_nl_ubl/__manifest__.py
@@ -1,15 +1,0 @@
-# Copyright 2020 Onestein (<https://www.onestein.eu>)
-# License AGPL-3.0 or later (https://www.gnu.org/licenses/agpl).
-
-{
-    "name": "Dutch - UBL",
-    "version": "13.0.1.0.0",
-    "category": "Dutch Localization",
-    "license": "AGPL-3",
-    "summary": "Adds Dutch-related data in UBL XML files",
-    "author": "Onestein, Odoo Community Association (OCA)",
-    "maintainers": ["astirpe"],
-    "website": "https://github.com/OCA/l10n-netherlands",
-    "depends": ["base_ubl"],
-    "installable": True,
-}
