--- l10n_nl_ubl/models/__init__.py
@@ -1,3 +1,0 @@
-# License AGPL-3.0 or later (https://www.gnu.org/licenses/agpl).
-
-from . import base_ubl
