--- l10n_nl_ubl/readme/CONTRIBUTORS.rst
@@ -1,1 +1,0 @@
-* Andrea Stirpe <a.stirpe@onestein.nl>
