--- l10n_nl_ubl/tests/__init__.py
@@ -1,4 +1,0 @@
-# License AGPL-3.0 or later (https://www.gnu.org/licenses/agpl).
-
-from . import test_ubl_post_install
-from . import test_ubl_at_install
